Add some fix from nightly-build

+++ src/shared/common/tables.h
@@ -405,8 +405,8 @@ typedef struct SPlayerTable
 	long	lExitX, lExitY;
 	long	lExitMapIndex;
 
-	short       hp;
-	short       sp;
+	int       hp;
+	int       sp;
 
 	short	sRandomHP;
 	short	sRandomSP;
